reorganize HARK.distributions
