+++ HARK/ConsumptionSaving/ConsPrefShockModel.py
@@ -22,7 +22,7 @@ from HARK.Calibration.Income.IncomeProcesses import (
     get_PermShkDstn_from_IncShkDstn,
     get_TranShkDstn_from_IncShkDstn,
 )
-from HARK.distribution import MeanOneLogNormal, expected
+from HARK.distributions import MeanOneLogNormal, expected
 from HARK.interpolation import (
     CubicInterp,
     LinearInterp,
