+++ HARK/core.py
@@ -20,7 +20,7 @@ from warnings import warn
 
 import numpy as np
 import pandas as pd
-from HARK.distribution import (
+from HARK.distributions import (
     Distribution,
     IndexDistribution,
     TimeVaryingDiscreteDistribution,
